+++ src/main/java/racing/domain/Car.java
@@ -1,4 +1,4 @@
-package step3.racing;
+package racing.domain;
 
 
 /**
